71713: Dropzone without uploader to own component
