+++ src/app/shared/file-dropzone-no-uploader/file-dropzone-no-uploader.scss
